requirements.txt
celery==5.4.0
fastapi==0.115.0
gunicorn==23.0.0
hypothesis==6.112.2
Markdown==3.7
pytest-cov==5.0.0
redis==5.1.0
requests==2.32.2
schemathesis==3.36.2
uvicorn[standard]==0.31.0